// src/app/modules/student/student.interface.ts
import { Model, Types } from 'mongoose';


export type TUserName = {
  firstName: string;
  middleName: string;
  lastName: string;
};

export type TGuardian = {
  fatherName: string;
  fatherOccupation: string;
  fatherContactNo: string;
  motherName: string;
  motherOccupation: string;
  motherContactNo: string;
};

export type TLocalGuardian = {
  name: string;
  occupation: string;
  contactNo: string;
  address: string;
};

export type TStudent = {
  id: string;
  user: Types.ObjectId;
  password: string;
  name: TUserName;
  gender: 'male' | 'female' | 'other';
  dateOfBirth?: Date;
  email: string;
  contactNo: string;
  emergencyContactNo: string;
  bloogGroup?: 'A+' | 'A-' | 'B+' | 'B-' | 'AB+' | 'AB-' | 'O+' | 'O-';
  presentAddress: string;
  permanentAddress: string;
  guardian: TGuardian;
  localGuardian: TLocalGuardian;
  profileImg?: string;
  admissionSemester: Types.ObjectId;
  isDeleted: boolean;
  academicDepartment: Types.ObjectId;
};

//for creating static

export interface StudentModel extends Model<TStudent> {
  isUserExists(id: string): Promise<TStudent | null>;
}

//----------------------------------------------------------------------------------------
// for creating instance

// export interface StudentMethods {
//   isUserExists(id: string): Promise<TStudent | null>;
// }

// export type StudentModel = Model<
//   TStudent,
//   Record<string, never>,
//   StudentMethods
// >;

// import { Model } from "mongoose";

// export type TUserName = {
//   firstName: string;
//   middleName?: string;
//   lastName: string;
// };

// export type TGuardian = {
//   fatherName: string;
//   fatherOccupation: string;
//   fatherContactNo: string;
//   motherName: string;
//   motherOccupation: string;
//   motherContactNo: string;
// };

// export type TLocalGuardian = {
//   name: string;
//   occuption: string;
//   contactNo: string;
//   address: string;
// };

// export type TStudent = {
//   id: string;
//   password: string;
//   name: TUserName;
//   gender: 'male' | 'female'|'other';
//   dateOfBrith?: string;
//   email: string;
//   contactNo: string;
//   emergencyContactNo: string;
//   bloodGroup?: 'A+' | 'A-' | 'B+' | 'B-' | 'AB+' | 'AB-' | 'O+' | 'O-';
//   presentAddress: string;
//   permanantAddress: string;
//   guardian: TGuardian;
//   localGuardian: TLocalGuardian;
//   profileImg?: string;
//   isActive: 'active' | 'blocked';
// };

// // for creating static

// export interface StudentModel extends Model<IStudent> {
//    isUserExist(id: string): Promise<TStudent | null>
// }

// // for creating instance
// // export type StudentMethods = {
// //   isUserExists(id: string): Promise<TStudent | null>;
// // };

// // export  type StudentModel = Model<TStudent, Record<string, never>, StudentMethods>;